webinit/requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dominate==2.3.1
greenlet==0.4.13
gunicorn==19.8.1
idna==2.6
inflection==0.3.1
itsdangerous==0.24
MarkupSafe==1.0
psycopg2-binary==2.7.4
PyMySQL==0.8.1
python-craigslist==1.0.6
python-dateutil==2.7.2
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
visitor==0.1.3
websocket-client==0.47.0

